refactor(settings): modularize schedules and todo runner tabs
