+++ src/renderer/components/settings/types.ts
@@ -1,0 +1,4 @@
+import type { SchedulerTask, TodoRunnerJob } from '../../types'
+
+export type SchedulerTaskDraft = SchedulerTask & { isDraft?: boolean }
+export type TodoRunnerJobDraft = TodoRunnerJob & { isDraft?: boolean; todoItemsText: string }
